update: install script

+++ install.sh
@@ -1,81 +1,114 @@
 #!/bin/bash
 
+# Store current stderr redirection
+exec 3>&2
+
 # Redirect stderr to /dev/null for the entire script
 exec 2>/dev/null
 
-# Get Project Directory
+#Get project directory
 project_dir="$(pwd)"
 
 declare -a packages=("git" "zsh" "wget" "php" "mysql" "node" "nvm" "composer" "lsd" "postgresql@13")
 declare -a applications=("google-chrome" "visual-studio-code" "warp" "raycast" "notion" "todoist" "postman" "dbeaver-community" "spotify")
 declare -a fonts=("font-jetbrains-mono" "font-meslo-lg-nerd-font")
 
-if ! command -v brew &> /dev/null; then
-    echo -e "Homebrew is not installed. Installing...\n"
+echo_message() {
+    echo -e "\n$1\n"
+}
+
+# Check and install Homebrew
+if ! command -v brew &>/dev/null; then
+    echo_message "Homebrew is not installed. Installing..."
     /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
-    echo 'eval "$(/opt/homebrew/bin/brew shellenv)"' >> ~/.zshrc
     eval "$(/opt/homebrew/bin/brew shellenv)"
 else
-    echo -e "Homebrew is already installed.\n"
+    echo_message "Homebrew is already installed."
 fi
 
+# Install required packages
 for package in "${packages[@]}"; do
-    if command -v "$package" &> /dev/null; then
-        echo -e "$package is already installed \n"
-    else
-        echo -e "Installing $package...\n"
+    if ! brew list --formula | grep -q "^$package\$"; then
+        echo_message "Installing $package..."
         brew install "$package" || { echo "Could not install packages" && exit; }
-        echo -e "$package Installed\n"
+        echo_message "$package installed."
+    else
+        echo_message "$package is already installed."
     fi
 done
 
+# Install applications
 for application in "${applications[@]}"; do
-    # Get cask info
     cask_info=$(brew info --cask "$application" 2>/dev/null)
-
-    # Extract the application name from the cask info
     application_name=$(echo "$cask_info" | awk '/Artifacts/{getline; sub(/ \(App\)/,""); sub(/^ */,""); print}')
-    if [ -n "$application_name" ]; then
-        echo -e "Found $application_name in Applications folder, skipping installation.\n"
+    if [ -z "$application_name" ] && ! brew list --cask "$application" &>/dev/null; then
+        brew install --cask "$application"
+        [ "$application" = "visual-studio-code" ] && export PATH="$PATH:/Applications/Visual Studio Code.app/Contents/Resources/app/bin"
+        echo_message "$application_name installed."
     else
-        if brew list --cask "$cask_name" &>/dev/null; then
-            echo -e "$application_name is already installed \n"
-        else
-            brew install --cask "$application"
-            [ "$application" = "visual-studio-code" ] && export PATH="$PATH:/Applications/Visual Studio Code.app/Contents/Resources/app/bin"
-            echo -e "$application_name installed\n"
-        fi
+        echo_message "$application_name is already installed."
     fi 
 done
 
-if brew tap | grep -q "homebrew/cask-fonts"; then
-    echo -e "homebrew/cask-fonts is already tapped \n"
-else
+# Tap cask fonts
+if ! brew tap | grep -q "homebrew/cask-fonts"; then
     brew tap homebrew/cask-fonts
 fi
 
+# Install fonts
 for font in "${fonts[@]}"; do
-    if brew list --cask | grep -q "^$font\$"; then
-        echo -e "$font is already installed \n"
-    else
+    if ! brew list --cask | grep -q "^$font\$"; then
         brew install --cask "$font"
+    else
+        echo_message "$font is already installed."
     fi 
 done
 
-# Backup original .zshrc file in the ~/.config/shazam/backup directory
-mkdir -p ~/.config/shazam/backup
-if mv -n ~/.zshrc ~/.config/shazam/backup/.$(date +"%Y%m%d")-zshrc-backup; then
-    echo -e "Backed up the current .zshrc to ~/.config/shazam/backup/.$(date +"%Y%m%d")-zshrc-backup \n"
+# Check MySQL service status
+mysql_service_status=$(brew services list | grep mysql | awk '{print $2}')
+
+# Install MySQL
+if [ "$mysql_service_status" != "started" ] && command -v mysql &>/dev/null; then
+    # Start MySQL service
+    brew services start mysql
+
+    # Secure MySQL installation
+    echo_message "Securing MySQL installation..."
+    # mysql_secure_installation
+
+    echo_message "MySQL setup complete."
+
+    # Ask the user if they want to restore databases
+    exec 2>&3 # Restore stderr redirection
+    read -r -p "Do you want to restore your databases? (yes/no): " restore_databases
+    exec 3>&- # Close file descriptor 3
+
+    if [[ $restore_databases =~ ^[Yy][Ee][Ss]$ ]]; then
+        read -r -p "Enter the file path for the database dump: " db_dump_file
+        if [ -f "$db_dump_file" ]; then
+            mysql -u root -p < "$db_dump_file"
+            echo_message "Database restored successfully."
+        else
+            echo_message "Error: File not found at $db_dump_file"
+        fi
+    fi
 fi
 
-echo -e "The setup will be installed in '~/.config/shazam' \n"
+# Backup original .zshrc file
+backup_dir="$HOME/.config/shazam/backup"
+mkdir -p "$backup_dir"
+if mv -n ~/.zshrc "$backup_dir/.$(date +"%Y%m%d")-zshrc-backup"; then
+    echo_message "Backed up the current .zshrc to $backup_dir."
+fi
 
-echo -e "Installing oh-my-zsh\n"
+echo_message "The setup will be installed in '~/.config/shazam'."
+
+echo_message "Installing oh-my-zsh."
 if [ -d ~/.config/shazam/oh-my-zsh ]; then
-    echo -e "oh-my-zsh is already installed \n"
+    echo_message "oh-my-zsh is already installed."
     git -C ~/.config/shazam/oh-my-zsh remote set-url origin https://github.com/ohmyzsh/ohmyzsh.git
 elif [ -d ~/.oh-my-zsh ]; then
-    echo -e "oh-my-zsh in already installed at '~/.oh-my-zsh'. Moving it to '~/.config/shazam/oh-my-zsh' \n"
+    echo_message "oh-my-zsh is already installed at '~/.oh-my-zsh'. Moving it to '~/.config/shazam/oh-my-zsh'."
     export ZSH="$HOME/.config/shazam/oh-my-zsh"
     mv ~/.oh-my-zsh ~/.config/shazam/oh-my-zsh
     git -C ~/.config/shazam/oh-my-zsh remote set-url origin https://github.com/ohmyzsh/ohmyzsh.git
@@ -83,88 +116,72 @@ else
     git clone --depth=1 https://github.com/ohmyzsh/ohmyzsh.git ~/.config/shazam/oh-my-zsh
 fi
 
-cp -f config/.aliases ~/.config/shazam/
-cp -f config/.shazam ~/.config/shazam/
-cp -f config/.p10k.zsh ~/.config/shazam/
-cp -f config/.zshrc ~/
-
-# this will be used to store .zcompdump zsh completion cache files which normally clutter $HOME
-mkdir -p ~/.cache/zsh/
-
-if [ -f ~/.zcompdump* ]; then
-    mv ~/.zcompdump* ~/.cache/zsh/
+# Create ~/.nvm directory if nvm is installed but the directory does not exist
+if command -v nvm &>/dev/null && [ ! -d "$HOME/.nvm" ]; then
+    mkdir -p "$HOME/.nvm"
 fi
 
-# Function to update or clone a git repository
+# Copy configuration files
+cp -f config/.aliases config/.shazam config/.p10k.zsh ~/.config/shazam/
+cp -f config/.zshrc ~/
+
+# Move .zcompdump files
+mkdir -p ~/.cache/zsh/
+[ -f ~/.zcompdump* ] && mv ~/.zcompdump* ~/.cache/zsh/
+
+# Update or clone git repositories
 update_or_clone_repository() {
     local repo_path="$1"
     local repo_url="$2"
     local repo_name=$(basename "$repo_path")
 
     if [ -d "$repo_path" ]; then
-        echo -e "Updating $repo_name...\n"
+        echo_message "Updating $repo_name..."
         cd "$repo_path" && git pull -q
     else
-        echo -e "Cloning $repo_name...\n"
+        echo_message "Cloning $repo_name..."
         git clone --depth=1 "$repo_url" "$repo_path"
     fi
 }
 
-# Update or clone zsh-autosuggestions plugin
 update_or_clone_repository "$HOME/.config/shazam/oh-my-zsh/plugins/zsh-autosuggestions" "https://github.com/zsh-users/zsh-autosuggestions"
-
-# Update or clone zsh-syntax-highlighting plugin
 update_or_clone_repository "$HOME/.config/shazam/oh-my-zsh/custom/plugins/zsh-syntax-highlighting" "https://github.com/zsh-users/zsh-syntax-highlighting.git"
-
-# Update or clone zsh-completions plugin
 update_or_clone_repository "$HOME/.config/shazam/oh-my-zsh/custom/plugins/zsh-completions" "https://github.com/zsh-users/zsh-completions"
-
-# Update or clone zsh-history-substring-search plugin
 update_or_clone_repository "$HOME/.config/shazam/oh-my-zsh/custom/plugins/zsh-history-substring-search" "https://github.com/zsh-users/zsh-history-substring-search"
-
-# Update or clone powerlevel10k theme
 update_or_clone_repository "$HOME/.config/shazam/oh-my-zsh/custom/themes/powerlevel10k" "https://github.com/romkatv/powerlevel10k.git"
 
-echo -e "All plugins and themes are up to date!\n"
+echo_message "All plugins and themes are up to date!"
 
-# Get extensions file
+# Install VS Code extensions
 extensions_file="$project_dir/vscode/.extensions"
 
-# Function to check if an extension is installed
 extension_installed() {
     local extension="$1"
     code --list-extensions | grep -q "$1" && return 0 || return 1
 }
 
-# Install extensions listed in the extensions file
 while IFS= read -r extension || [[ -n "$extension" ]]; do
-    if extension_installed "$extension"; then
-        echo -e "Extension '$extension' is already installed. \n"
-    else
-        echo -e "Installing extension: $extension"
+    if ! extension_installed "$extension"; then
+        echo_message "Installing extension: $extension"
         code --install-extension "$extension"
+    else
+        echo_message "$extension already installed"
     fi
 done < "$extensions_file"
 
-# Path to custom settings.json file
+# Copy custom VS Code settings
 custom_settings_file="$project_dir/vscode/settings.json"
-
-# Path to VS Code settings directory
 vscode_settings_dir="$HOME/Library/Application Support/Code/User"
-
-# Check if VS Code settings directory exists and copy settings if it does
-if [ ! -d "$vscode_settings_dir" ]; then
-    echo -e "Error: VS Code settings directory does not exist: $vscode_settings_dir \n"
-else
-    cp "$custom_settings_file" "$vscode_settings_dir/settings.json"
-    echo -e "Copied VS Code settings.json \n"
+if [ -d "$vscode_settings_dir" ]; then
+    cp -f "$custom_settings_file" "$vscode_settings_dir"
+    echo_message "Copied VS Code settings.json"
 fi
 
-# Update oh-my-zsh using Zsh
+# Update oh-my-zsh
 if /bin/zsh -c 'source ~/.zshrc && omz update'; then
-    echo -e "\nInstallation Complete!"
+    echo_message "\nInstallation Complete!"
 else
-    echo -e "\nSomething is wrong"
+    echo_message "\nSomething went wrong."
 fi
 
 exit
